libs/voice-pipeline/examples/webapp-webrtc/requirements.txt
# WebRTC Backend Requirements
aiortc>=1.6.0
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
msgpack>=1.0.0
aiohttp>=3.9.0
numpy>=1.24.0
python-dotenv>=1.0.0
websockets>=12.0

# Voice Pipeline Dependencies
torch>=2.0.0
torchaudio>=2.0.0
faster-whisper>=1.0.0
anthropic>=0.25.0
kokoro>=0.9.0
soundfile>=0.12.0

# Voice Pipeline (local install)
# voice-pipeline @ file:../..
